Do nothing when no shell is defined, just exit.
I was thinking how I could make it so that build steps could be
executed, without running any process. It turns out, that even
without this change, a Command without shell can be defined
and used to just trigger a set of build steps.
This change only prevents the application from scraping together
the arguments, which would then be never used anways.
Maybe this reduces the time from 140 ms to something lower :D