Replace Command.capture_all with Process args template [[argv]]
This way static args and command line input can be combined for command execution.
Also, the argv can be placed anywhere inside the args array (as its own string element),
it is then expanded with the whole input.
It is NOT turned into a string, the args array is modified!