Split Hisho into Hisho Core and Hisho CLI
Hisho Core only provides the functionality of Hisho as a standalone library.
Hisho CLI uses Hisho Core to implement the command line utility.

This split has been made so that the functionality of Hisho can be integrated
into other applications, and to allow multiple frontends of Hisho to exist.

For example, we could now add a new Clap based CLI with different command
structure or a GUI application while still providing the same functionality.

As of this split, all features are still available as they where before.