Core: Implement initial support for "explain"
I had to create some functions that have their arity in their name,
this had to be done because Rust does not allow for argument overloading.
It is not possible to have two functions with the same name take different
typed arguments.
So I appended the Arity, which is the count of arguments the function takes.